5.3D/requirements.txt
PyQt5
RPi.GPIO